src/game_generator/pyproject.toml
[tool.poetry]
name = "game-generator"
version = "0.1.0"
description = ""
authors = ["guilherme096 <guilherme.rosa60@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
